renamed tabsfolder to bundleEditor
