+++ src/app/options/outputs/OutputsPage.tsx
@@ -9,7 +9,7 @@ import { Output } from "@/model/Output";
 import { BundleSettingsHandler } from "@/utils/BundleSettingsHandler";
 import { invoke } from "@tauri-apps/api/core";
 import { useEffect, useState } from "react";
-import TabFolders from "./components/TabFolders";
+import BundleEditor from "./components/BundleEditor";
 
 const OutputsPage = () => {
   const [editBundle, setEditBundle] = useState<Bundle | undefined>(undefined); // [1
@@ -151,7 +151,7 @@ const OutputsPage = () => {
             </div>
           )}
           {editMode && outputs.length > 0 && (
-            <TabFolders
+            <BundleEditor
               outputs={outputs}
               toggleOutput={toggleOutput}
               dialogOpen={dialogOpen}
