+++ src/components/elements/inputs/Input.tsx
@@ -1,8 +1,8 @@
 import { ReactElement } from "react";
-import InfoWindow, { InfoWindowProps } from "../InfoWindow";
-import { Checkbox } from "./inputs/Checkbox";
-import { TextArea } from "./inputs/TextArea";
-import { Label } from "./Label";
+import { InfoWindowProps } from "../../InfoWindow";
+import { Checkbox } from "./Checkbox";
+import { TextArea } from "./TextArea";
+import { Label } from "../Label";
 
 interface InputProps {
   label?: string;
